Backend/requirements.txt
python-dotenv
google-genai
feedparser
diffusers
transformers
accelerate
torch
Pillow